Do 50,000 post burn-in MCMC iterations, keeping 2,000 of them

+++ R/sim-study-2/functions.R
@@ -99,8 +99,8 @@ run_mcmc <- function(data, design, voom=TRUE){
                             alpha = init_run[['state']]$alpha,
                             C = C)
   
-  contr$n_iter <- as.integer(200000)
-  contr$thin <- as.integer(50)
+  contr$n_iter <- as.integer(50000)
+  contr$thin <- as.integer(25)
   contr$warmup <- as.integer(10000)
   contr$idx_save <- 0:(dat$G-1)
   contr$n_save_P <- as.integer(100)
